Delete attributes.h/compat.h, more
