+++ Common/Common.vcxproj
@@ -390,33 +390,28 @@
     <ClInclude Include="..\ext\at3_standalone\atrac3data.h" />
     <ClInclude Include="..\ext\at3_standalone\atrac3plus.h" />
     <ClInclude Include="..\ext\at3_standalone\atrac3plus_data.h" />
-    <ClInclude Include="..\ext\at3_standalone\attributes.h" />
     <ClInclude Include="..\ext\at3_standalone\avcodec.h" />
     <ClInclude Include="..\ext\at3_standalone\avfft.h" />
-    <ClInclude Include="..\ext\at3_standalone\bswap.h" />
     <ClInclude Include="..\ext\at3_standalone\bytestream.h" />
     <ClInclude Include="..\ext\at3_standalone\channel_layout.h" />
     <ClInclude Include="..\ext\at3_standalone\common.h" />
     <ClInclude Include="..\ext\at3_standalone\compat.h" />
-    <ClInclude Include="..\ext\at3_standalone\config.h" />
     <ClInclude Include="..\ext\at3_standalone\error.h" />
     <ClInclude Include="..\ext\at3_standalone\fft-internal.h" />
     <ClInclude Include="..\ext\at3_standalone\fft.h" />
     <ClInclude Include="..\ext\at3_standalone\float_dsp.h" />
     <ClInclude Include="..\ext\at3_standalone\get_bits.h" />
-    <ClInclude Include="..\ext\at3_standalone\internal.h" />
     <ClInclude Include="..\ext\at3_standalone\intfloat.h" />
     <ClInclude Include="..\ext\at3_standalone\intmath.h" />
     <ClInclude Include="..\ext\at3_standalone\intreadwrite.h" />
     <ClInclude Include="..\ext\at3_standalone\libm.h" />
     <ClInclude Include="..\ext\at3_standalone\log.h" />
-    <ClInclude Include="..\ext\at3_standalone\macros.h" />
     <ClInclude Include="..\ext\at3_standalone\mathematics.h" />
-    <ClInclude Include="..\ext\at3_standalone\mathops.h" />
     <ClInclude Include="..\ext\at3_standalone\mem.h" />
-    <ClInclude Include="..\ext\at3_standalone\rational.h" />
+    <ClInclude Include="..\ext\at3_standalone\qsort.h" />
     <ClInclude Include="..\ext\at3_standalone\samplefmt.h" />
     <ClInclude Include="..\ext\at3_standalone\sinewin.h" />
+    <ClInclude Include="..\ext\at3_standalone\sinewin_tablegen.h" />
     <ClInclude Include="..\ext\at3_standalone\util_internal.h" />
     <ClInclude Include="..\ext\at3_standalone\version.h" />
     <ClInclude Include="..\ext\basis_universal\basisu.h" />
@@ -623,7 +618,6 @@
     <ClCompile Include="..\ext\at3_standalone\atrac3plus.c" />
     <ClCompile Include="..\ext\at3_standalone\atrac3plusdec.c" />
     <ClCompile Include="..\ext\at3_standalone\atrac3plusdsp.c" />
-    <ClCompile Include="..\ext\at3_standalone\avfft.c" />
     <ClCompile Include="..\ext\at3_standalone\bitstream.c" />
     <ClCompile Include="..\ext\at3_standalone\channel_layout.c" />
     <ClCompile Include="..\ext\at3_standalone\compat.c" />
@@ -633,8 +627,7 @@
     <ClCompile Include="..\ext\at3_standalone\mathematics.c" />
     <ClCompile Include="..\ext\at3_standalone\mdct_template.c" />
     <ClCompile Include="..\ext\at3_standalone\mem.c" />
-    <ClCompile Include="..\ext\at3_standalone\options.c" />
-    <ClCompile Include="..\ext\at3_standalone\rational.c" />
+    <ClCompile Include="..\ext\at3_standalone\avcodec.c" />
     <ClCompile Include="..\ext\at3_standalone\samplefmt.c" />
     <ClCompile Include="..\ext\at3_standalone\sinewin.c" />
     <ClCompile Include="..\ext\at3_standalone\utils.c" />
