+++ ext/at3_standalone/avutil.h
@@ -188,9 +188,7 @@
 
 #include "common.h"
 #include "error.h"
-#include "rational.h"
 #include "version.h"
-#include "macros.h"
 #include "mathematics.h"
 #include "log.h"
 
@@ -223,17 +221,6 @@ unsigned av_int_list_length_for_size(unsigned elsize,
 #define av_int_list_length(list, term) \
     av_int_list_length_for_size(sizeof(*(list)), list, term)
 
-/**
- * Open a file using a UTF-8 filename.
- * The API of this function matches POSIX fopen(), errors are returned through
- * errno.
- */
-FILE *av_fopen_utf8(const char *path, const char *mode);
-
-/**
- * Return the fractional representation of the internal time base.
- */
-AVRational av_get_time_base_q(void);
 
 /**
  * @}
