+++ ext/at3_standalone/bitstream.c
@@ -28,21 +28,15 @@
  * bitstream api.
  */
 
+#include <stdlib.h>
+#include <string.h>
+
 #include "qsort.h"
 #include "avcodec.h"
-#include "internal.h"
+#include "common.h"
 #include "util_internal.h"
-#include "mathops.h"
+#include "mathematics.h"
 #include "get_bits.h"
-
-const uint8_t ff_log2_run[41]={
- 0, 0, 0, 0, 1, 1, 1, 1,
- 2, 2, 2, 2, 3, 3, 3, 3,
- 4, 4, 5, 5, 6, 6, 7, 7,
- 8, 9,10,11,12,13,14,15,
-16,17,18,19,20,21,22,23,
-24,
-};
 
 /* VLC decoding */
 
@@ -61,7 +55,6 @@ const uint8_t ff_log2_run[41]={
         break;                                              \
     }                                                       \
 }
-
 
 static int alloc_table(VLC *vlc, int size, int use_static)
 {
