--- ext/at3_standalone/config.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "compat.h"
