+++ ext/at3_standalone/float_dsp.c
@@ -19,8 +19,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#include "config.h"
-#include "attributes.h"
+#include "compat.h"
 #include "float_dsp.h"
 #include "mem.h"
