+++ ext/at3_standalone/libm.h
@@ -26,8 +26,7 @@
 #define AVUTIL_LIBM_H
 
 #include <math.h>
-#include "config.h"
-#include "attributes.h"
+#include "compat.h"
 #include "intfloat.h"
 #include "mathematics.h"
