+++ ext/at3_standalone/sinewin.h
@@ -21,7 +21,6 @@
 #ifndef AVCODEC_SINEWIN_H
 #define AVCODEC_SINEWIN_H
 
-#include "config.h"
 #include "mem.h"
 #include "aac_defines.h"
