+++ ext/at3_standalone/utils.c
@@ -26,15 +26,13 @@
  */
 
 #include "util_internal.h"
-#include "config.h"
 #include "avcodec.h"
-#include "attributes.h"
+#include "compat.h"
 #include "channel_layout.h"
-#include "internal.h"
+#include "common.h"
 #include "mathematics.h"
 #include "samplefmt.h"
 #include "avcodec.h"
-#include "internal.h"
 #include "bytestream.h"
 #include "version.h"
 #include <stdlib.h>
@@ -42,60 +40,8 @@
 #include <limits.h>
 #include <float.h>
 
-static int (*lockmgr_cb)(void **mutex, enum AVLockOp op) = NULL;
-
-
-volatile int ff_avcodec_locked;
-static int volatile entangled_thread_counter = 0;
-static void *codec_mutex;
-static void *avformat_mutex;
-
-void av_fast_padded_malloc(void *ptr, unsigned int *size, size_t min_size)
-{
-    uint8_t **p = ptr;
-    if (min_size > SIZE_MAX - AV_INPUT_BUFFER_PADDING_SIZE) {
-        av_freep(p);
-        *size = 0;
-        return;
-    }
-    if (!ff_fast_malloc(p, size, min_size + AV_INPUT_BUFFER_PADDING_SIZE, 1))
-        memset(*p + min_size, 0, AV_INPUT_BUFFER_PADDING_SIZE);
-}
-
-static av_cold void avcodec_init(void)
-{
-    static int initialized = 0;
-
-    if (initialized != 0)
-        return;
-    initialized = 1;
-}
-
-int av_codec_is_decoder(const AVCodec *codec)
-{
-    return 1;
-}
-
-#if FF_API_EMU_EDGE
-unsigned avcodec_get_edge_width(void)
-{
-    return EDGE_WIDTH;
-}
-#endif
-
-int attribute_align_arg ff_codec_open2_recursive(AVCodecContext *avctx, const AVCodec *codec, void **options)
-{
-    int ret = 0;
-
-    ret = avcodec_open2(avctx, codec, options);
-
-    return ret;
-}
-
 int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *codec, void **options)
 {
-    int ret = 0;
-
     if ((codec && avctx->codec && codec != avctx->codec)) {
         av_log(avctx, AV_LOG_ERROR, "This AVCodecContext was allocated for %s, "
                                     "but %s passed to avcodec_open2()\n", avctx->codec->name, codec->name);
@@ -106,6 +52,8 @@ int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *code
 
     if (avctx->extradata_size < 0 || avctx->extradata_size >= FF_MAX_EXTRADATA_SIZE)
         return AVERROR(EINVAL);
+
+	int ret = 0;
 
     if (codec->priv_data_size > 0) {
         if (!avctx->priv_data) {
@@ -132,13 +80,6 @@ int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *code
    
     avctx->frame_number = 0;
 
-#if FF_API_VISMV
-    if (avctx->debug_mv)
-        av_log(avctx, AV_LOG_WARNING, "The 'vismv' option is deprecated, "
-               "see the codecview filter instead.\n");
-#endif
-
-
     if (avctx->codec->init) {
         ret = avctx->codec->init(avctx);
         if (ret < 0) {
@@ -148,36 +89,24 @@ int attribute_align_arg avcodec_open2(AVCodecContext *avctx, const AVCodec *code
 
     ret=0;
 
-#if FF_API_AUDIOENC_DELAY
-    if (av_codec_is_encoder(avctx->codec))
-        avctx->delay = avctx->initial_padding;
-#endif
-
-    if (av_codec_is_decoder(avctx->codec)) {
-        /* validate channel layout from the decoder */
-        if (avctx->channel_layout) {
-            int channels = av_get_channel_layout_nb_channels(avctx->channel_layout);
-            if (!avctx->channels)
-                avctx->channels = channels;
-            else if (channels != avctx->channels) {
-				char buf[512] = "";
-                av_log(avctx, AV_LOG_WARNING,
-                       "Channel layout '%s' with %d channels does not match specified number of channels %d: "
-                       "ignoring specified channel layout\n",
-                       buf, channels, avctx->channels);
-                avctx->channel_layout = 0;
-            }
+    /* validate channel layout from the decoder */
+    if (avctx->channel_layout) {
+        int channels = av_get_channel_layout_nb_channels(avctx->channel_layout);
+        if (!avctx->channels)
+            avctx->channels = channels;
+        else if (channels != avctx->channels) {
+			char buf[512] = "";
+            av_log(avctx, AV_LOG_WARNING,
+                    "Channel layout '%s' with %d channels does not match specified number of channels %d: "
+                    "ignoring specified channel layout\n",
+                    buf, channels, avctx->channels);
+            avctx->channel_layout = 0;
         }
-        if (avctx->channels && avctx->channels < 0 ||
-            avctx->channels > FF_SANE_NB_CHANNELS) {
-            ret = AVERROR(EINVAL);
-            goto free_and_end;
-        }
-
-#if FF_API_AVCTX_TIMEBASE
-        if (avctx->framerate.num > 0 && avctx->framerate.den > 0)
-            avctx->time_base = av_inv_q(av_mul_q(avctx->framerate, (AVRational){avctx->ticks_per_frame, 1}));
-#endif
+    }
+    if (avctx->channels && avctx->channels < 0 ||
+        avctx->channels > FF_SANE_NB_CHANNELS) {
+        ret = AVERROR(EINVAL);
+        goto free_and_end;
     }
     if (codec->priv_data_size > 0 && avctx->priv_data && codec->priv_class) {
         av_assert0(*(const AVClass **)avctx->priv_data == codec->priv_class);
@@ -189,12 +118,6 @@ end:
 free_and_end:
     if (avctx->codec)
         avctx->codec->close(avctx);
-
-#if FF_API_CODED_FRAME
-FF_DISABLE_DEPRECATION_WARNINGS
-    av_frame_free(&avctx->coded_frame);
-FF_ENABLE_DEPRECATION_WARNINGS
-#endif
 
     av_freep(&avctx->priv_data);
     avctx->codec = NULL;
